+++ src/main/java/chess/ChessApplication.java
@@ -1,6 +1,5 @@
 package chess;
 
-import chess.domain.game.ChessGame;
 import chess.view.InputView;
 import chess.view.OutputView;
